+++ saleor/graphql/order/mutations/order_fulfill.py
@@ -116,6 +116,10 @@ class OrderFulfill(BaseMutation):
                 description="Order is fulfilled.",
             ),
             WebhookEventInfo(
+                type=WebhookEventAsyncType.FULFILLMENT_TRACKING_NUMBER_UPDATED,
+                description="Sent when fulfillment tracking number is updated.",
+            ),
+            WebhookEventInfo(
                 type=WebhookEventAsyncType.FULFILLMENT_APPROVED,
                 description="A fulfillment is approved.",
             ),